restruturação do código
