--- errorstatus.h
@@ -1,5 +1,0 @@
-#ifndef ERRORSTATUS_H
-#define ERRORSTATUS_H
-#define SUCESSO 1
-#define FALHA 0
-#endif // ERRORSTATUS_H
